Add new authorisation feature and test user detail
We are intending to update the seeding of the database for non-production environments to include a set of known users. See [Update seed script to create test users](https://github.com/DEFRA/sroc-tcm-admin/pull/305).

As soon as this is merged and available we will have a set of known users with specific roles and access to set regimes. With this knowledge we can create some tests to confirm the authorisation functionality is behaving as expected.

This also serves as a nice example of using the updated user config and includes a Cucumber data table based scenario as a 'Brucie bonus'.