Fix transaction feature for all environments
Until we are happy to use these tests instead of the one in [sroc-tcm-acceptance-tests](https://github.com/DEFRA/sroc-tcm-acceptance-tests) we need to protect the existing data in the environments. This is because the old test is dependent on it existing.

Once we have reached a point where we can be confident in just using this project, then we can follow our normal practise of being free to reset and seed the non-production environments at will, in order to support testing.

So our current issue is we do not have consistent data across the environments. So, to fix this we are making the customer reference to search for an environment configuration item. In this way we can get the transactions feature working across all environments.