Fix transaction test
Actually, the problem was that in some environments the `system` user is setup differently. This means the regimes it has access to is limited, or it was used to access a regime other than `CFD` and this has defaulted when the tests have run. So, we've made 2 key changes

- switched to a user we know is setup the same across all 3 environments
- updated the test to ensure the right regime is selected before searching

Knowing this we were able to identify a customer that exists across all 3 environments. With these changes plus some going away and creating all the test users across all non-prod environments, all features are now passing everywhere!