Update to match features in CHA accp. test project
We recently added the project [SROC Charging Module API tests](https://github.com/DEFRA/sroc-charging-module-api-tests) to our suite of tools. It is also Cypress based and will be used to test the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) going forward.

We started with this project as a base and then made the necessary amendments to allow us to test an API. Along the way we also made a few other improvements. This change is about bringing those improvements back into this project and keeping the 2 consistent to make it easier to switch between them.

The keys changes are

- use `*.env` files for all customer configuration and store them in an `/environments` folder
- add the ability to generate a test report to make it easier to share results with other stakeholders