Update dependencies and locations
We update our dependency versions to match the API project. We also ensure the dependencies are in the right place.

Because we ran `npm install` with the latest version of npm it's bumped the version from 1 to 2. Finally, we fixed those vulnerabilities that could be done using `npm audit fix`.