Add an .nvmrc file / Delete config
An .nvmrc file is a handy way for those new to a project to see what version of Node is being used. Most of us are also [nvm](https://github.com/nvm-sh/nvm) users which means it can use the file to automatically select the right version of Node from those installed.

Finally, it'll support CI for the project which we'll be adding in a later commit.

----

We delete the config folder and .env.example because we are moving to the way the API tests do things, which is using .env solely and the term 'environments'.