Switch to environments
Using these environment files instead of config do pretty much the same thing as the old `config/` folder. The key difference is we now drive everything through env vars rather than having a mix of Cypress config and dotenv environment variables.

This made things simpler in the API tests and making the same change should make switching between the 2 easier for those that use both projects because the behaviour will be the same.