Add CI to the project
Because our services are behind a VPN we can't actually test if our 'tests' work when we push changes. But we can at least confirm we haven't broken the project with our change.

This CI feature is a common one we use for this purpose and it has appeared in some form across a range of acceptance test projects for Defra!