First pass at upload function
We initially tried putting our upload code in `common/import.js` using AWS SDK v3. We ran into problems importing v3 so we reverted to v2. This at least imported successfully but we were unable to upload files -- it would throw a network error when we tried it. We thought this may be due to `import.js` being run in the Cypress browser so we moved the code into a task `s3Upload` as this would be run by the underlying Node server. And this worked! The next stage will be to see if v3 works in a task, and whether there's a better way of structuring our task (eg. moving it to its own file). But at least we now have a working solution as a base.