Update aws-sdk to v3
When we were trying to perform the upload in `import.js` we initially had problems with aws-sdk v3 so we reverted to using v2. Once we realised that we had to do the upload in a task, we continued using v2 and developed a working solution. We now replace this with v3 of the sdk.