Add current test import files
These are the files we currently use when we need to both import some data and test the transaction import process. We still need to automate the process of getting them imported. But adding them to `/features` at least means they have a home and will start to have a change history.