Add legacy CFD feature from Java project
https://eaflood.atlassian.net/browse/CMEA-273

The [SROC TCM Acceptance tests](https://github.com/DEFRA/sroc-tcm-acceptance-tests) is actually a single Selenium-based test that goes through the TCM checking it for each of the 3 regimes. It was what the current team were handed and is currently the most extensive test of the TCM service.

However, it is impossible to maintain and hard to follow what the test is actually doing. What we do know of it is that there are few assertions. The test assumes if it can click through with no errors the TCM is functioning correctly. But we have had examples where this is not the case.

This change migrates what the legacy test is doing for the Installations (PAS) regime into our Cypress framework. Re-writing the test in BDD format should make it easier to understand what actions are being taken. From there we can review the test, breaking it down into valuable features we can maintain going forward.