Add email feature
https://eaflood.atlassian.net/browse/CMEA-289

We recently [Added the ability to read last email for testing](https://github.com/DEFRA/sroc-tcm-admin/pull/533) to the TCM. With that in place we can now automate the manual testing we currently do to check email is still working during release sign off.

This change adds a new feature that covers various email scenarios.