Add 2 new page objects
This covers the create user account page and the set password page. When a new account is created an email is sent with a link to the set password page. The new user is expected to follow the link and then set a new password for themselves.

We need these page objects in place before we can build a passing journey.