Add LastEmailPage
This is not a true page object in that there is no UI to interact with. But it is an endpoint in the TCM and the only one of its kind so we add it to our page objects.

The logic, unlike the other page objects, is focused on polling the page to obtain the right email details and then extracting the link we'll need to use in order to accept the invite.