Add faker-js package
We need to generate new users, and specifically email addresses that are unique to avoid the email test failing because the user already exists.

We could just generate some random value but why not have something that better represents what would actually be entered? This might also prove useful in future for generating fake data.