Add new helpers.js module to support
As the comment in the file explains we sometimes have logic in our steps that is 'vanilla' JS and could be used in other steps. We can reduce the complexity and noise by moving the functionality here.

It also means we have a place we can add these types of functions in the future.