Add new custom commands
There was some repitition in our steps for the email feature. Plus, we were being forced to sign in as one kind of user, then sign out and complete the test as another.

This is because only admin users can create other users but we want to run our tests as a basic user.

So, we moved some of this reusable functionality into custom commands. We also where we could switched from going through the UI and following the full journey to do things like sign in and add user in certain cases. It's not what the test is actually tryting to cover and therefore can make the test output very noisy.

Most acceptance test suites will encourage you to avoid always going through the UI for things like signing in when what you are actually trying to test isa some other unrelated feature.

Because of Rails CSRF protections you must have a valid csrf-token when you `POST` the request through. Other than building a completly new API only endpoint the best approximation we can get is visiting the page in question, extracting a CRSF token, and then including it in our request.