Add a BasePage and extend all from it
We start by creating a `BasePage` which we then extend all other pages from. It takes over duties for the `mainHeading()`. We also move to some conventions.

- elements will feature their type in the name, for example, `cancelButton`
- rather than refer to the main submit button on each page by its display name we go with one that is consistent. This also means we can declare it once in our `BasePage`
- we expect all pages to provide a `confirm()` method. This will replace checks on `mainHeading()` littered throughout the test to confirm we're in the right place. Our convention now will be to ask the page object to confirm we're where we are supposed to be

These changes mean a number of edits to the steps. Along the way whilst retesting we found some broken elements. So, scattered throughout the changes will be ones related to fixing a test rather than the switch to a base page.