Add BaseAppPage
The base app page represents what you see once you have authenticated. This means the various menus. The intention is to access the menus through the page object that represents where you are rather than referencing menus separartly in the steps. We create a base class for it to remove duplication. We want to try and get referencing them out of the steps because we think it will help make them cleaner and clearer.