Add fixturePickerHelper()
Before we can start importing the test file for a regime we need a way to determine which file to use.

We could have had the step in the feature just state the file name. But we felt it would read nicer if we instead stated the regime and then have something determine what file to use from that, for example, "Given I am starting with known cfd data".