Add killSession() command
We are going to need to trigger the `/jobs/import` endpoint in the TCM to get out test transaction files to import. But you can only access it as an admin user.

We already have a command that allows us authenticate without going through the UI. What we don't have is a command that then allows us to sign out (kill the session).