Increase wait()
Grrrr! Now for 'reasons' we don't understand this started failing. The issue was Cypress was claiming an element was covering the checkbox it was trying to `check`.

We already had a wait() in there which was dealing with it. But recent runs has required us to increase it to prevent errors.