Add a common step to clean and import regime data
We actually figured this out when working on automating this aspect of the legacy tests. But because it looks like we'll have to update those tests to account for these issues we're seeing it made sense to get the bugs committed first and then go back and make our changes to the legacy tests.