Add common step to reset all regimes
When first working on the bugs we started with the idea of resetting the system for all regimes rather than one at a time. As it turns out we could reproduce the bugs just using one regime. So, to save time we reverted to just resetting the system for that regime.

But should be ever add a bug that applies to a different regime than installations (and because it took awhile to get this step working!) we're adding it now so it's available for use in the future.