Add commands to support new data reset steps
We are going to need to trigger the `/jobs/import` endpoint in the TCM to get out test transaction files to import. But you can only access it as an admin user. We already have a command that allows us authenticate without going through the UI. What we did't have is a command that then allows us to sign out (kill the session).

Next we need hit the TCM's `/jobs/import` endpoint. As we're going to need to hit `/jobs/export` in future as well we have added this new `runJob()` command that gives us the ability to do both.

Finalyy, we cleaned up some of the comments and the order that the commands appear.