Move Download data checks to own feature
https://eaflood.atlassian.net/browse/CMEA-299

This ia a first step in automating the test of the download transaction data function. We currently have some steps in the legacy tests that just check we can get to the page and see the heading.

That's not actually doing anything. So, taking it out of the legacy tests is no great loss and helps simplify them. But it also gives us something we can expand once we've added support to the TCM to trigger generating the data export files.