Fic the formatting in the file
We're using the VS .Code extension [Cucumber (Gherkin) Full Support](https://marketplace.visualstudio.com/items?itemName=alexkrechik.cucumberautocomplete) when it comes to writing the features. It includes a formatter we've been using to keep the files consistent.

We've run the formatter against this feature to bring the formatting into line.