Move exclude transaction to own feature
All 3 legacy tests include steps where we exclude then immediately reinstate a transaction. The steps as they are don't really do much but do make the lgecay tests more complex than they need to be.

This change moves them to their own feature and expands the steps to include a bit more checking of what we expect to happen when a transaction is excluded.