Add our first table object
A large bulk of our steps involve interacting with tables in the UI. The tables show the results of searches and are our primary source of extracting data we then need in subsequent steps. Prior to this change all this interaction was happening in the steps. How we referred to a table, its rows and columns wasn't always consistent and involved a lot of duplication.

We can though treat these tables in the same way we treat our menus and pages; as reusable objects in our steps. So, this change adds our first table object (which just happens to be the primary one) `TransactionsToBeBilledTable`. We've created a `BaseTable` following the pattern set down for page objects then extended our transactions to be billed table from it.

We were always a bit unsure of `columnPickerHelper()` because though strongly tied to the transactions table we don't make that explicit. Now we have a table object we can move that code to where it arguably makes more sense.

We expect to add more table objects in the future.