Update transactions general feature
The transactions general feature was making use of the old table methods so we needed to update it. Whilst in there we realised we could make use of our new data import step to ensure the expected that was present. Finally, we could simplify the scenario by logging in with the user that matches the data we've just imported.