Move legacy page export steps to new feature
This is following the same pattern as

- [Move legacy transaction history check to feature](https://github.com/DEFRA/sroc-acceptance-tests/pull/80)
- [Move legacy sort checks to their own feature](https://github.com/DEFRA/sroc-acceptance-tests/pull/76)
- [Move exclude transaction to own feature](https://github.com/DEFRA/sroc-acceptance-tests/pull/78)

In the legacy tests when we get to some pages (not all!) we hit the page's export button. A feature of the tcm is its ability to download the currently filtered data to a CSV file. To clean up and simplify the legacy tests we'll move these checks to their own feature.