Move legacy page checking steps to new feature
This is following the same pattern as

- [Move legacy transaction history check to feature](https://github.com/DEFRA/sroc-acceptance-tests/pull/80)
- [Move legacy sort checks to their own feature](https://github.com/DEFRA/sroc-acceptance-tests/pull/76)
- [Move exclude transaction to own feature](https://github.com/DEFRA/sroc-acceptance-tests/pull/78)
- [Move legacy page export steps to new feature](https://github.com/DEFRA/sroc-acceptance-tests/pull/82)
- [Move legacy transaction detail steps to new feature](https://github.com/DEFRA/sroc-acceptance-tests/pull/83)

The legacy tests don't do anything more than check that if you go to a page the title is as expected. But it does add noise to the key thing the legacy tests are checking; updating a transaction to the point it can be exported for invoicing.

So, we move them to their own feature.