Rename existing change password page
Turns out there are 2 change password pages in the TCM; one you get to in the 'forgot password' journey and the other when you click 'Change Password' once logged in.

We want to check we can access the latter but the ideal name was already taken. So, this change renames the existing page to make room for the new page we want to add.