Add a version of splitData() to helpers file
We're trying to centralise all our standard javascript functions in one place (helpers.js). This way our steps contain only Cypress calls and we're better able to reuse the functions we add.

We're going to need this method in the lgeacy tests so ahead of that we add a version to `helpers.js`.