Update the legacy WML feature
We're focused on this one first as it only involves generating a standard transaction file (no pre-sroc).

One of the first changes is we need to grab both the customer reference and the permit (consent) reference as the permit reference does not get exported. Whilst making the change we take advantage of our new page object table functionality.

Then we add a couple of new steps; one to kick of the export job in the TCM (normally it would be done on a 10 minute schedule) and the other to actually download the file and make assertions against it.