Add option to override authorisation
In the WML leagcy feature authenticating as the admin and then killing the session was not an issue because the steps to check the export file are at the end of the feature.

But the CFD and WML features have additional steps after the first export. When we ran the previous version of the command they failed because the session had been killed.

So, we have added an option to the command to allow us to override the authentication related calls. By default it will work as before but now we can tell it not to bother logging in before running the job, and killing the session after. As we are already logged in as the admin within the legacy steps we can use the current session to visit the endpoints.