Add support for testing pre-sroc trans files
CFD and WML also generate a pre-sroc file. We initially tried to reuse the steps but then hit an issue in that CFD has multiple customer references being exported in the pre-sroc file.

So, we amended the step. We stopped trying to be clever and grabbing the value to check from the UI. Instead we fall back to presumption we know what's in our test data and therefore what we expect to see in the export file. This means using a Cucumber data table and passing the expected values through.

It also means we no longer need to capture the customer reference.